fix: Finalizing branch for merge.
- Use u128 intermediates in all power rate calculations: types.rs
  from_ast piecewise path, config/units.rs nj_per_timestep, and
  parse_flow_rate now returns None on u128->u64 overflow.
- Fix unbounded f.read() in modules_lora_mesh and modules_mixed_network
  examples (same FUSE EOF bug already fixed in modules_solar_sensor).
- Update doc/modules.md ESP32 example to match corrected stdlib values.
- Use tracing::warn! instead of eprintln for poisoned mutex warning.
- Make CLI category filter fully case-insensitive (prefix check too).
- Add test for case-insensitive profile lookup during merge.
- Revert scrollbar drag direction to original sign convention.

+++ config/src/units.rs
@@ -182,9 +182,9 @@ impl PowerRate {
     ///
     /// Formula: energy_nj = rate_nw × timestep_ns / time_ns
     pub fn nj_per_timestep(&self, timestep_ns: u64) -> u64 {
-        let rate_nw = self.rate * self.unit.to_nw_factor();
-        let time_ns = self.time.to_ns_factor();
-        rate_nw * timestep_ns / time_ns
+        let rate_nw = self.rate as u128 * self.unit.to_nw_factor() as u128;
+        let time_ns = self.time.to_ns_factor() as u128;
+        (rate_nw * timestep_ns as u128 / time_ns) as u64
     }
 }
